+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-beta.2] 2022-03-20
+### Changed
+- Class name is now a required second parameter
+- README
+
 ## [2.0.0-beta.1] 2022-03-20
 ### Changed
-- Changed element to be a required first parameter
+- Element is now a required first parameter
 - Moved types to lib
 
 ## [1.2.5] - 2022-03-20